requirements.txt
git+https://github.com/twrecked/pyaarlo
python-decouple
python-statemachine
